# docs/api/openapi.yml
openapi: 3.0.0
info:
  title: Defcon API
  description: Defcon control plane API documentation
  version: 0.1.0
...
paths:
  /api/checks:
    get:
      summary: List checks
      description: By default, this endpoint lists all enabled checks defined in the system. If all is provided with a `1` value, list **all** checks (enabled or not).
      tags: ['Checks']
      parameters:
        - name: all
          description: Should we display all checks or enabled checks only
          in: query
          schema:
            type: boolean
            enum: [false, true]
            default: false
        - name: group
          description: UUID of the group of which to return the checks
          in: query
          schema:
            type: string
            format: uuid
        - name: kind
          description: Slug for the handler types to return
          in: query
          schema:
            type: string
        - name: site
          description: Name of the site for which to return the checks
          schema:
            type: string
      responses:
        '200':
          description: List of enabled checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/specs/ReadSpec'

    post:
      summary: Create a new check
      tags: ['Checks']
      requestBody:
        description: The definition of a check
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/specs/WriteSpec'
      responses:
        '201':
          description: The check was created
          headers:
            Location:
              description: The absolute path to the check, including the unique identifier
              schema:
                type: string
                format: url
        '400':
          description: Invalid payload
  /api/checks/{uuid}:
    get:
      summary: Retrieve a check
      tags: ['Checks']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
          required: true
      responses:
        '200':
          description: The definition of the check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/specs/ReadSpec'
        '404':
          description: Provided UUID does not match a known check

    put:
      summary: Update a check
      description: Update a check by providing a full valid definition. Note that the handler type (`spec.kind`) cannot be updated in-place.
      tags: ['Checks']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        description: The definition of a check
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/specs/WriteSpec'
      responses:
        '200':
          description: The check was updated
        '400':
          description: Invalid payload
        '404':
          description: Provided UUID does not match a known check

    patch:
      summary: Update check attributes
      description: Update some of the attributes of a check through a partial definition. Note that the handler type (`spec.kind`) cannot be updated in-place and that you have to specify the full handler definition (`spec`) if you wish to update it.
      tags: ['Checks']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        description: The partial definition of a check
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/check/PatchWriteCheck'
      responses:
        '200':
          description: The check was updated
        '400':
          description: Invalid payload
        '404':
          description: Provided UUID does not match a known check

    delete:
      summary: Disable or delete a check
      description: Defcon does not allow to completely delete a check, only to disable it (prevent it from running).
      tags: ['Checks']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - name: delete
          description: Delete the check instead of disabling it
          in: query
          schema:
            type: boolean
            enum: [false, true]
            default: false
      responses:
        '204':
          description: Check was disabled or deleted successfully
        '404':
          description: Provided UUID does not match a known check
...